+++ Problem03/main.cpp
@@ -14,16 +14,16 @@ namespace {
 
 int main() {
     aoc::InputReader input{};
-    int total_uno{0};
+    long total_uno{0};
     const auto &input_lines = input.getLines();
     auto is_gear_symbol = [](char c) {
         return c == '*';
     };
-    std::map<const char *, std::vector<int>> gear_candidates{};
-    for (int i = 0; i < input_lines.size(); ++i) {
+    std::map<const char *, std::vector<long>> gear_candidates{};
+    for (long i = 0; i < input_lines.size(); ++i) {
         const auto &prev_line = input_lines[std::max(0, i - 1)];
         const auto &current_line = input_lines[i];
-        const auto &next_line = input_lines[std::min(i + 1, static_cast<int>(input_lines.size() - 1))];
+        const auto &next_line = input_lines[std::min(i + 1, static_cast<long>(input_lines.size() - 1))];
         auto number_start = std::find_if(current_line.begin(), current_line.end(),
                                          [](char c) { return std::isdigit(c); });
         while (number_start != current_line.end()) {
@@ -35,23 +35,23 @@ int main() {
             auto prev_line_symbol = std::find_if(prev_line.begin() + min_index, prev_line.begin() + max_index + 1, is_gear_symbol);
             if(&prev_line != &current_line && prev_line.begin() + max_index + 1 != prev_line_symbol ) {
                 gear_candidates[
-                        prev_line_symbol.base()].push_back(aoc::getInt({number_start, number_end}));
+                        prev_line_symbol.base()].push_back(aoc::getLong({number_start, number_end}));
             }
             auto next_line_symbol = std::find_if(next_line.begin() + min_index, next_line.begin() + max_index + 1, is_gear_symbol);
             if(&next_line != &current_line &&
                next_line.begin() + max_index + 1 != next_line_symbol) {
                 gear_candidates[
-                        next_line_symbol.base()].push_back(aoc::getInt({number_start, number_end}));
+                        next_line_symbol.base()].push_back(aoc::getLong({number_start, number_end}));
 
             }
             if ( is_gear_symbol(current_line[min_index])) {
                 gear_candidates[
-                        current_line.c_str()+min_index].push_back(aoc::getInt({number_start, number_end}));
+                        current_line.c_str()+min_index].push_back(aoc::getLong({number_start, number_end}));
 
             }
             if ( is_gear_symbol(current_line[max_index])) {
                 gear_candidates[
-                        current_line.c_str()+max_index].push_back(aoc::getInt({number_start, number_end}));
+                        current_line.c_str()+max_index].push_back(aoc::getLong({number_start, number_end}));
 
             }
 
@@ -59,7 +59,7 @@ int main() {
         }
     }
 
-    int total{0};
+    long total{0};
     for(const auto &candidate : gear_candidates) {
         if(candidate.second.size() == 2) {
             total += candidate.second[0] * candidate.second[1];
